Move loadUser from WichtelUserService to AuthService
