+++ backend/src/main/java/org/example/backend/service/WichtelUserService.java
@@ -22,7 +22,7 @@ import static org.example.backend.util.DTOConverter.fromDTO;
 
 @Service
 @RequiredArgsConstructor
-public class WichtelUserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
+public class WichtelUserService{
 
     private final WichtelUserRepo repo;
     private final IdService idService;
@@ -50,33 +50,5 @@ public class WichtelUserService implements OAuth2UserService<OAuth2UserRequest, 
         return repo.findAll();
     }
 
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2UserService<OAuth2UserRequest, OAuth2User> delegate = new DefaultOAuth2UserService();
-        OAuth2User oAuth2User = delegate.loadUser(userRequest);
 
-        String provider = userRequest.getClientRegistration().getRegistrationId();
-        Map<String,Object> attr = oAuth2User.getAttributes();
-        String providerId = attr.get("id").toString();
-
-        String email = requireNonNullElse(attr.get("email"),"").toString();
-        String name = requireNonNullElse(attr.get("login"),"").toString();
-
-        Optional<WichtelUser> userOptional = repo.findByOauthProviderAndOauthId(provider,providerId);
-        if(userOptional.isEmpty()) {
-            WichtelUser user = WichtelUser.builder()
-                    .email(email)
-                    .name(name)
-                    .oauthProvider(provider)
-                    .oauthId(providerId)
-                    .build();
-            repo.save(user);
-        }
-
-        return new DefaultOAuth2User(
-                null,
-                oAuth2User.getAttributes(),
-                "id"
-        );
-    }
 }
